requirements.txt
python-pptx>=0.6.21
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=1.0.0
google-search-results>=2.4.2
uvicorn[standard]
fastapi 